refactor: move Flag enum to its own module
